requirements.txt
fastapi
uvicorn[standard]

openai
ollama
mcp

sqlalchemy
psycopg[binary]
alembic
pgvector

pydantic
pydantic-settings
python-dotenv

httpx
msal
msgraph-sdk
hubspot-api-client
google-api-python-client

apscheduler

python-multipart
jinja2

tenacity
orjson